commit app_barcode_scanner_widget.dart auth.jpg auth_screen.dart bar_code_screen_check.dart check_sheet_products.dart choose_store_screen.dart GeneratedPluginRegistrant.swift hero_dialog_route.dart main.dart pubspec.lock pubspec.yaml route_settings.dart screens.dart widgets.dart

+++ lib/config/route_settings.dart
@@ -34,9 +34,9 @@ class RouteSettingsWithArguments extends RouteSettings {
       case CheckSheetDetailScreen.routeName:
         return MaterialPageRoute(
             builder: (_) => CheckSheetDetailScreen(
-                  product: (settings.arguments as Map<String, dynamic>)['product']
-                      as ProductDTO,
-                  index: (settings.arguments as Map<String, dynamic>)['index'],
+                  product: (settings.arguments
+                      as Map<String, dynamic>)['product'] as ProductDTO,
+                  index: (settings.arguments as Map<String, dynamic>)['index'] as int,
                 ));
 
       default:
